+++ src/sourcepawn_preprocessor/src/lib.rs
@@ -1,4 +1,5 @@
 pub(crate) mod evaluator;
+mod macros;
 pub mod preprocessor;
 mod preprocessor_operator;
